Remove topics specific code
